requirements.txt
streamlit
fastapi
uvicorn
requests
pandas
numpy
scikit-learn
python-dotenv
mysql-connector-python
plotly
openpyxl
tensorflow

